Issue 1093 Stage table actions (#1110)
* quick login for developers on studio

* fix reactivity not working

* update stage inside studio

* update status & visibility mutation

* table visibility & status toggle

+++ studio/src/utils/common.ts
@@ -10,8 +10,12 @@ export function getSharedConfig(): SharedConfigs {
     const sharedConfig = JSON.parse(localStorage.getItem("configs") ?? "");
     return sharedConfig;
   } catch (error) {
-    localStorage.clear(); // Remove shared auth so that it will ask you to visit Dashboard for login
-    return {} as SharedConfigs;
+    if (import.meta.env.PROD) {
+      localStorage.clear(); // Remove shared auth so that it will ask you to visit Dashboard for login
+    }
+    return {
+      GRAPHQL_ENDPOINT: import.meta.env.VITE_GRAPHQL_ENDPOINT,
+    } as SharedConfigs;
   }
 }
